In Test3Widget, calculate file ast independently (only when source changes).
This is an "optimization", or doing things as I actually "intended" by
recalculating the ast only when it changes, rather than every time it
is needed. Unfortunately, the current way to make it work adds a lot of
code bloat and makes it harder to understand/maintain/change. :( Gotta
find/create a better way.
What used to be 3 lines in (*Test3Widget) NotifyChange(), becomes 4
lines in (*parsedFile) NotifyChange(), plus 8 lines for parsedFile
struct, plus 1 line in Test3Widget, plus 2 lines in NewTest3Widget().