Use ViewGroup for MultilineContentI, fixes unnecessary file writes.
ViewGroup implements DepNode2ManualI.
This is a good beginning of a proper solution for multiple
views/editors into the same content. It keeps a track of all connected
views, and when one is updated, it issues updates only to all others.
That also allows for adding/removing views during runtime, and having
more than 2 connected views.