Add LayoutNeeded() to Widgeter.
LayoutNeeded(), previously planned as Layout2(), is called in the first
pass on all visible widgets to perform layout as needed. This way, the
second pass of Render() on all visible widgets is only concerned with
drawing within the calculated laid out space and does not do layout.
Add content Widgeter to WindowWidget.