vcs: Do not interactively ask for a password (for git).
Fixes potential "Username for 'https://github.com': " interactive
prompts for private repos.
Thank you @pornel for the solution from bower/bower#1009. It took me
half an hour to find a working solution for this problem, which is odd
given that ls-remote is supposed to be a low-level plumbing command to
be used primarily by scripts. :/