Use a separate experimental package for serving Markdown with automatic Table of Contents.
I am still looking for a good solution to the problem, but until then
this should work and allow Conception-go to compile and work. It relies
on a hardcoded relative path, so the current solution is not general
and doesn't allow other packages to import u10. That's why I didn't
want to commit it to the higher quality general markdown_http package
all this time.
Depends on shurcooL/go@aee0b312a664dfef96ef717e4e1cdd6de6a915ad.
Fixes #6.