Deprecate experimental project to shurcooL-legacy namespace.
This is not an actively developed project anymore. It was experimental
in nature, and I've used it to explore and prototype many ideas. I'm
spending my time elsewhere now. Since the code here is old, written in
haste with the goal of testing prototypes rather than long term
maintenance, it's quite old and of low quality.

I'm using my github.com/shurcooL namespace for active, high quality Go
packages only, so move this project from github.com/shurcooL to
shurcooL-legacy to better reflect the state it's in.

If there's desire to continue some of the work here, it will be
extracted bit by bit and re-created as high quality Go packages.

Closes #3.
Closes #4.