add go.mod file, start building in module mode
When using the GOPATH mode, unless one vendors all dependencies or
uses third-party dependency management tools, there is no viable
choice but to always automatically opt-in to latest versions of
all dependencies.

That is actually desirable behavior for most of my actively developed
and actively maintained projects, because it means I don't have to do
extra work of checking for newer versions of dependencies and updating
to them when they come out; that happens automatically and implicitly.
It helps detect breakages as soon as possible and keeps the overall
maintenance costs down. It has the acceptable to me cost of the project
being potentially broken for some short amount of time (but well within
my 95%+ yearly uptime SLA stated at https://github.com/shurcooL/SLA).

However, that behavior means a new version of a dependency may come
out that breaks the build, and I'd have to spend time on fixing
Conception-go. Since Conception-go is a legacy project, that is not
wanted.

Solve this problem by switching to using module mode for building
Conception-go. It produces reproducible builds and doesn't
automatically opt-in to newer versions of dependencies.

Note, however, that Conception-go itself still expects Go packages
that it operates on to be available in a GOPATH workspace.
Updating Conception-go to be compatible with Go's module mode is
not currently planned (because it's a frozen legacy project).