place shurcooL/frontend repository in GOPATH workspace
It's still needed by GopherJS (which hasn't been updated to support
module mode yet) and due to issue https://golang.org/issue/31603.

Modify BuildPackageFromImportPath function to provide the current
working directory to build.Import, so that it can work successfully
when Conception-go has been installed in module mode and its assets
(alongside its source code) are not available in the GOPATH workspace.

cd $HOME before running GO111MODULE=on go install in order to prevent
it from modifying the current module, if the user happens to be inside
a module when they run the installation command.

Update Linux install instructions to use the current Go version.