update dependencies
Take advantage of a recent update to the GopherJS compiler that
allows it to build with newer versions of Go to make it so that
Conception-go can be built with the latest version of Go.

Also update to glfw/v3.3, since it fixes warnings and improves
compatibility.

Updates https://github.com/gopherjs/gopherjs/issues/941.